+++ examples/resources/morpheus_app_blueprint_catalog_item/import.sh
@@ -1,0 +1,1 @@
+terraform import morpheus_app_blueprint_catalog_item.tf_example_app_blueprint_catalog_item 1