+++ examples/resources/morpheus_instance_catalog_item/resource.tf
@@ -1,8 +1,10 @@
 resource "morpheus_instance_catalog_item" "tf_example_instance_catalog_item" {
   name        = "tfexample_instance_catalog"
   description = "terraform example instance catalog item"
+  image_path  = "tfexample.png"
+  image_name  = "tfexample.png"
   enabled     = true
-  feature     = true
+  featured    = true
   content     = <<TFEOF
   {"name":"test"}
   TFEOF